// src/app/models/line-tank-type.ts
export class LineTankType {
    lineTankTypeId: number;
    lineId: number;
    lineName: string;
    tankTypeId: number;
    tankTypeName: string;
    sequence: number;
    createdBy: string;
    updatedBy: string;
    datetimeCreated: number;
    datetimeUpdated: number;

    highlighted = false;
}